#100 Merge `tests/end2end/test_public_forgot_password.py` into `tests/end2end/test_public_reset_password.py`
